application/ai/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cpu

annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.12
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
filelock==3.18.0
    # via torch
fsspec==2025.3.0
    # via torch
google-api-core==2.24.2
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.1.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.69.2
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
hydra-core==1.3.2
    # via sam2
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iopath==0.1.10
    # via sam2
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.4
    # via
    #   -r requirements.in
    #   opencv-python-headless
    #   sam2
    #   shapely
    #   torchvision
omegaconf==2.3.0
    # via hydra-core
opencv-python-headless==4.11.0.86
    # via -r requirements.in
packaging==24.2
    # via hydra-core
pdf2image==1.17.0
    # via -r requirements.in
pillow==11.1.0
    # via
    #   pdf2image
    #   sam2
    #   torchvision
portalocker==3.1.1
    # via iopath
proto-plus==1.26.1
    # via google-api-core
protobuf==6.30.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.1
    # via fastapi
pydantic-core==2.33.0
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   omegaconf
    #   uvicorn
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.1
    # via fastapi-cli
rsa==4.9
    # via google-auth
sam2==1.1.0
    # via -r requirements.in
shapely==2.0.7
    # via -r requirements.in
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
starlette==0.46.1
    # via fastapi
sympy==1.13.1
    # via torch
torch==2.6.0
    # via
    #   -r requirements.in
    #   sam2
    #   torchvision
torchvision==0.21.0
    # via
    #   -r requirements.in
    #   sam2
tqdm==4.67.1
    # via
    #   iopath
    #   sam2
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.13.0
    # via
    #   anyio
    #   fastapi
    #   iopath
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
